+++ Documents/Help/Changelog.md
@@ -1,6 +1,11 @@
 # <img src="https://github.com/Krypton-Suite/Standard-Toolkit-Demos/blob/master/Krypton.png?raw=true"> Standard Toolkit Demos - ChangeLog
 
-## 2023-11-## - Build 2311 - November 2023
+## 2024-11-xx - Build 2311 - November 2024
+* Add `KryptonAboutToolkit` example
+* Removed support for .NET 7
+* Add support for .NET 9
+
+## 2023-11-14 - Build 2311 - November 2023
 * Relayout in alphabetical order the Compnents view in the Launch Explorer
 * Add missing Examples to launcher
 * Addition of new example for Progress Bars